fk it init push

+++ common/buildcraft/lib/client/sprite/SpriteAtlas.java
@@ -24,11 +24,11 @@ public class SpriteAtlas implements ISprite {
 
     @Override
     public double getInterpU(double u) {
-        return sprite.getU(u * 16);
+        return sprite.getU((float) (u * 16));
     }
 
     @Override
     public double getInterpV(double v) {
-        return sprite.getV(v * 16);
+        return sprite.getV((float) (v * 16));
     }
 }
